rename CredentialsContextProvider and remove unused import
